Safari-611.1.10/LayoutTests/media/media-constants-expected.txt

Test HTMLMediaElement and MediaError constants.

EXPECTED (HTMLMediaElement.NETWORK_EMPTY == '0') OK
EXPECTED (HTMLMediaElement.NETWORK_IDLE == '1') OK
EXPECTED (HTMLMediaElement.NETWORK_LOADING == '2') OK
EXPECTED (HTMLMediaElement.NETWORK_NO_SOURCE == '3') OK

EXPECTED (HTMLMediaElement.HAVE_NOTHING == '0') OK
EXPECTED (HTMLMediaElement.HAVE_METADATA == '1') OK
EXPECTED (HTMLMediaElement.HAVE_CURRENT_DATA == '2') OK
EXPECTED (HTMLMediaElement.HAVE_FUTURE_DATA == '3') OK
EXPECTED (HTMLMediaElement.HAVE_ENOUGH_DATA == '4') OK

EXPECTED (MediaError.MEDIA_ERR_ABORTED == '1') OK
EXPECTED (MediaError.MEDIA_ERR_NETWORK == '2') OK
EXPECTED (MediaError.MEDIA_ERR_DECODE == '3') OK
EXPECTED (MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED == '4') OK
END OF TEST
